refactor outbound packages
