+++ factory/service.go
@@ -12,8 +12,8 @@ import (
 	"hpc-express-service/gcs"
 	inbound "hpc-express-service/inbound/express"
 	"hpc-express-service/mawb"
-	cargoManifest "hpc-express-service/outbound/cargoManifest"
-	draftMawb "hpc-express-service/outbound/draftMawb"
+	cargoManifest "hpc-express-service/outbound/cargomanifest"
+	draftMawb "hpc-express-service/outbound/draftmawb"
 	outboundExpress "hpc-express-service/outbound/express"
 	outboundMawb "hpc-express-service/outbound/mawb"
 	"hpc-express-service/outbound/mawbinfo"
