+++ outbound/draftmawb/draft_mawb.go
@@ -1,4 +1,4 @@
-package outbound
+package draftmawb
 
 import (
 	"net/http"
