+++ outbound/draftmawb/service.go
@@ -1,4 +1,4 @@
-package outbound
+package draftmawb
 
 import (
 	"context"
