+++ server/backup_mawbinfo/mawbinfo.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	cargoManifest "hpc-express-service/outbound/cargoManifest"
-	draftMawb "hpc-express-service/outbound/draftMawb"
+	cargoManifest "hpc-express-service/outbound/cargomanifest"
+	draftMawb "hpc-express-service/outbound/draftmawb"
 	"hpc-express-service/setting"
 	"log"
 	"net/http"
