+++ server/mawbinfo.go
@@ -4,8 +4,8 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	cargoManifest "hpc-express-service/outbound/cargoManifest"
-	draftMawb "hpc-express-service/outbound/draftMawb"
+	cargoManifest "hpc-express-service/outbound/cargomanifest"
+	draftMawb "hpc-express-service/outbound/draftmawb"
 	weightslip "hpc-express-service/outbound/weightSlip"
 	"hpc-express-service/setting"
 	"log"
